+++ src/memory/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod error;
 
 #[cfg(any(test, feature = "bench"))]
-pub(crate) mod in_memory_store;
+pub(crate) mod in_memory;
 
 #[cfg(feature = "redis-mem")]
-pub(crate) mod redis_store;
+pub(crate) mod redis;
